+++ table/src/main/java/tech/ydb/table/rpc/grpc/GrpcTableRpc.java
@@ -15,7 +15,6 @@ import tech.ydb.core.grpc.GrpcReadStream;
 import tech.ydb.core.grpc.GrpcRequestSettings;
 import tech.ydb.core.grpc.GrpcTransport;
 import tech.ydb.core.operation.OperationManager;
-import tech.ydb.proto.StatusCodesProtos.StatusIds;
 import tech.ydb.proto.table.YdbTable;
 import tech.ydb.proto.table.YdbTable.AlterTableRequest;
 import tech.ydb.proto.table.YdbTable.AlterTableResponse;