+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.2.0 ##
 * Added export module
 * Operation manager
+* Added readRows operation in table service
 
 ## 2.1.2 ##